modified consumer file to test connection to AWS. TS issues
