+++ kafka/consumer.ts
@@ -1,12 +1,19 @@
 import * as Kafka from 'kafkajs';
-import ddb_putitem from '../database/ddb_putitem';
+import AWS from 'aws-sdk';
 
+AWS.config.update({
+  region: "us-east-2",
+  // endpoint: "http://localhost:8000"
+});
 /**
  * Creates a producer object for export later
  * @param groupId - required for consumer creation in Kafka
  * @param topic - topic that matches producer
  *
  */
+
+const ddb = new AWS.DynamoDB();
+
 const createConsumer = async (groupId: string, topic: string) => {
   // Creating a new Kafka instance
   // Configure AWS IAM client with kafka broker
@@ -32,7 +39,7 @@ const createConsumer = async (groupId: string, topic: string) => {
   console.log('something');
   await consumer.run({
     // autoCommitInterval: 5000,
-    // eachMessage: awsConnector.sendEachMessage, -> to send to the DB
+    // eachMessage: sendEachMessage, -> to send to the DB
     eachMessage: async ({ topic, partition, message }) => {
       console.log(message);
       try {
@@ -50,13 +57,61 @@ const createConsumer = async (groupId: string, topic: string) => {
       }
     },
   });
+}
 
-  // Error Handler for DynamoDB data insertion
-  // function insertHandler(err) {
-  //   if (err) {
-  //     console.log(`Error caught in DynamoDB database insertion" ${err}`);
-  //     }
-  //   }
-};
+// add code to insert data into dynamoDB
+// type sendEachMessage = () => void;
 
-export { createConsumer };
+const sendEachMessage = async () => {
+  // Declare a new variable set to the parsed JSON of message value
+  // const messageObj = await JSON.parse(message.value);
+  // console.log("The message recieved inside sendEachMessage is : ", messageObj);
+
+  //name of the dynamoDB table to which data is being inserted
+  const tableName = "patientData";
+
+  // destructure the object to send the data to dynamo DB
+  const params = {
+    TableName: tableName,
+    Item: { 
+      'eventId' : {S: '1234'}
+      // JSON(messageObj)
+    // 'eventId': {S: message.value.eventId },
+    // 'eventDateTime': {S: message.value.eventDateTime}, 
+    // 'eventName': {S: message.value.eventName} ,
+    // 'firstName': {S: message.value.firstName}, 
+    // 'lastName': {S: message.value.lastName},
+    // 'middleName': {S: message.value.middleName},
+    // 'gender': {S: message.value.gender},
+    // 'street': {S: message.value.street},
+    // 'street2': {S: message.value.street2},
+    // 'city': {S: message.value.city},
+    // 'county': {S: message.value.county},
+    // 'state': {S: message.value.state},
+    // 'zip': {N: message.value.zip},
+    // 'latitude': {N: message.value.latitude},
+    // 'longitude': {N: message.value.longitude}, 
+    // 'phone': {N: message.value.phone},
+    // 'email': {S: message.value.email}, 
+    // 'jobArea': {S: message.value.jobArea}, 
+    // 'jobTitle': {S: message.value.jobTitle},
+    // 'nameLastContact': {S: message.value.nameLastContact},
+    // 'dateLastContact': {S: message.value.dateLastContact},
+    // 'countryLastTravel': {S: message.value.countryLastTravel},
+    // 'dateLastTravel': {S: message.value.dateLastTravel}
+  }
+}
+
+  console.log("adding a new item from consumer to AWS DB..")
+
+// Send/put data to dynamoDB from the consumer
+ddb.putItem(params, function(err,data) {
+  if(err) {
+      console.error("Unable to add item, Error: ", err);
+  }
+  else {
+      console.log("Added Item : ", JSON.stringify(data) )
+  }
+});
+}
+  export { createConsumer };
